requirements.txt
pandas
python-dotenv
python-binance
